+++ packages/components/src/DataList/components/DataListFilters/DataListFilters.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { act, cleanup, render, screen } from "@testing-library/react";
 import { configMocks, mockIntersectionObserver } from "jsdom-testing-mocks";
 import { DataListFilters, InternalDataListFilters } from "./DataListFilters";
-import { CONTAINER_TEST_ID } from "./components/DataListOverflowFade/DataListOverflowFade.const";
+import { CONTAINER_TEST_ID } from "../DataListOverflowFade/DataListOverflowFade.const";
 import { defaultValues } from "../../context/DataListContext";
 import * as dataListContext from "../../context/DataListContext/DataListContext";
 import * as useShowHeader from "../../hooks/useShowHeader";
